Use 64-bit counts in the configurator (fix overflow at corpus scale)
On Windows `long` is 32-bit, so the accumulating counts wrapped on a large corpus:
the 43GB FineWeb run reported "alpha words total: -1184634348" (3.1B wrapped). Widen
the count types to 64-bit -- TokStats, word_freq, the BPE pair counts (pc), the
name-stat maps, names_withheld and quote_repl -- so totals and per-pair counts stay
correct (a frequent pair can exceed 2^31 on a corpus larger than FineWeb, which would
otherwise corrupt the merge selection). normalize_text's per-chunk `replaced` stays
32-bit (bounded by a chunk, and shared with the engine).

Byte-identical on the 256MB slice (values < 2^31 there); the word-count report is now
correct.

+++ include/sub0/casing.hpp
@@ -44,10 +44,10 @@ inline bool is_word_byte(int s) {
 
 // Per-corpus truecasing statistics (configurator reporting only).
 struct TokStats {
-    long words = 0;  // alpha runs seen
-    long cap   = 0;  // collapsed to <|cap|> + lowercase
-    long up    = 0;  // collapsed to <|up|>  + lowercase
-    long names = 0;  // capitalized/upper words left verbatim (corpus-aware)
+    long long words = 0;  // alpha runs seen
+    long long cap   = 0;  // collapsed to <|cap|> + lowercase
+    long long up    = 0;  // collapsed to <|up|>  + lowercase
+    long long names = 0;  // capitalized/upper words left verbatim (corpus-aware) (64-bit: a large corpus has >2^31 words)
 };
 
 // Fold "fancy" typographic glyphs to their ASCII equivalents. The corpus carries
